requirements-ray.txt
grpcio==1.48.2
ray[rllib]>=2.3,<2.6
munch
pyarrow>=7.0
numpy<1.24
gymnasium
pandas
seaborn
matplotlib
setuptools
tensorboard
hydra-core
optuna
tensorflow-probability
wandb
onnx
fsspec
gputil
plotly
jupyter

